+++ frontend/src/components/Reviews/UserSingleReview.js
@@ -17,7 +17,7 @@ function UserSingleReview({review}) {
   }
 
   const deleteHandler = async () => {
-    await dispatch(removeReview(review));
+    if (window.confirm('Are you sure you want to delete this review?')) await dispatch(removeReview(review));
   }
 
   useEffect(() => {
